# expression: implement vectorized evaluation for `builtinRpadBinarySig`