PCP #12106
<!--
Thank you for contributing to TiDB! Please read TiDB's [CONTRIBUTING](https://github.com/pingcap/community/blob/master/CONTRIBUTING.md) document **BEFORE** filing this PR.
-->
### What problem does this PR solve? <!--add issue link with summary if exists-->
implement vectorized evaluation for builtinRpadBinarySig, for #12106
### What is changed and how it works?
```
goos: linux
goarch: amd64
pkg: github.com/pingcap/tidb/expression
BenchmarkVectorizedBuiltinStringFunc/builtinRpadBinarySig-VecBuiltinFunc-4                   364           3095730 ns/op          450560 B/op       6144 allocs/op
BenchmarkVectorizedBuiltinStringFunc/builtinRpadBinarySig-NonVecBuiltinFunc-4                535           2204419 ns/op          450560 B/op       6144 allocs/op
BenchmarkVectorizedBuiltinStringFunc/builtinRpadBinarySig-VecBuiltinFunc#01-4                603           1970974 ns/op          289960 B/op       3954 allocs/op
BenchmarkVectorizedBuiltinStringFunc/builtinRpadBinarySig-NonVecBuiltinFunc#01-4             836           1453098 ns/op          289960 B/op       3954 allocs/op
[2019/11/10 15:40:06.939 +08:00] [WARN] [builtin_string.go:268] ["unexpected `Flen` value(-1) in CONCAT's args"] ["arg's index"=0]
[2019/11/10 15:40:06.939 +08:00] [WARN] [builtin_string.go:268] ["unexpected `Flen` value(-1) in CONCAT's args"] ["arg's index"=1]
[2019/11/10 15:40:06.939 +08:00] [WARN] [builtin_string.go:268] ["unexpected `Flen` value(-1) in CONCAT's args"] ["arg's index"=2]
PASS
ok      github.com/pingcap/tidb/expression      5.767s

```
### Check List <!--REMOVE the items that are not applicable-->
Tests <!-- At least one of them must be included. -->
- Unit test